% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_to.R
\name{assign_to}
\alias{assign_to}
\title{Assign to a name after a chain of data manipulations}
\usage{
assign_to(x, name, where = NULL)
}
\arguments{
\item{x}{a value to be assigned to \code{name}.}

\item{name}{an unquoted variable name.}

\item{where}{where to do the assignment. Default is the environment in which
the chain of manipulations is called. Passed to \code{pos} argument of
\code{\link{assign}}.}
}
\value{
The function is invoked for its side effect, which is to assign
  \code{x} to the variable \code{name}. \code{x} is then returned.
}
\description{
Alias for the base function \code{\link{assign}}, which makes it easier to
assign the output of a chain of data manipulations.
}
\examples{
mtcars \%>\%
  subset(select = c("cyl","drat", "wt", "qsec")) \%>\%
  by(.$cyl, colMeans) \%>\%
  assign_to(car_cyls)

print(car_cyls)

# It is also possible to assign in the middle of a chain.
mtcars \%>\%
  subset(select = c("cyl","drat", "wt", "qsec")) \%>\%
  by(.$cyl, colMeans) \%>\%
  assign_to(car_cyls) \%>\%
  lapply(function(x) as.data.frame(t(x))) \%>\%
  set_names(NULL) \%>\%
  do.call("rbind", .) \%>\%
  assign_to(carl_cyls_df)
  ls()

print(car_cyls)
print(carl_cyls_df)

}

