% Generated by roxygen2 (4.0.1): do not edit by hand
\name{\%<\%}
\alias{\%<\%}
\title{backpipe}
\usage{
lhs \%<\% rhs
}
\arguments{
\item{lhs}{a function or expression}

\item{rhs}{value ro be piped

There is not yet much support for complex expressions}
}
\description{
Simlilar to \code{\%>\%} but reverses the order of the operators. It is
especially useful for composing shiny tags.
}
\examples{
letters \%>\% paste0( 1:26 )  # forward pipe
paste0( 1:26 ) \%<\% letters  # backward pipe

\dontrun{
  div( class="outer-outer") \%<\%
    div( class="outer") \%<\%
      div( class="inner") \%<\% h1( "content", role="heading" )
}
}
\author{
Christopher Brown
}
\seealso{
\code{\link{\%<\%}}
}

