% Generated by roxygen2 (4.0.2): do not edit by hand
% Please edit documentation in R/getters.R
\name{calling_env}
\alias{calling_env}
\title{Get the calling environment of the pipe chain}
\usage{
calling_env()
}
\value{
If called inside a pipe chain, the calling environment of
  the pipe chain. Otherwise, the parent frame.
}
\description{
This function extracts the original environment from which the pipe
chain was called.
}
\details{
\code{calling_env()} must be called at the top level of the pipe
chain component.
}
\seealso{
\code{\link{parent.frame}()}
}

