% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.composite}
\alias{print.composite}
\title{Print method for composite functions.}
\usage{
\method{print}{composite}(x, ...)
}
\arguments{
\item{x}{a composite function}

\item{...}{not used.}
}
\description{
Generic method for printing of composite functions generated with
either \code{\%o>\%} or \code{compose}.
}

