% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%<$>\%}
\alias{\%<$>\%}
\title{magrittr compound assignment with exposition pipe-operator}
\usage{
lhs \%<$>\% rhs
}
\arguments{
\item{lhs}{An object which serves both as the initial value and as target.}

\item{rhs}{An expression where the names in lhs is available.}
}
\description{
Pipe an object forward into a function or call expression while exposing the names in \code{lhs} to the \code{rhs} expression,
then update the \code{lhs} object with the resulting value.
}
\details{
The compound assignment with exposition pipe-operator, \code{\%<$>\%}, is a
combination of the compound assignment pipe-operator, \code{\%<>\%}, and the exposition
pipe-operator, \code{\%$\%}, and it used to update a value by first piping into one or more rhs expressions,
while exposing its contents, and then assigning the result.
For example, \code{some_object \%<$>\%
foo \%>\% bar} is equivalent to \code{some_object <- some_object \%$\% foo
\%>\% bar}. It must be the first pipe-operator in a chain, but otherwise it
works like \code{\link{\%$\%}}.
}
\examples{
d <- data.frame(x=c(1,2), y=c(2,3))
d \%<$>\% rbind(., data.frame(x=sum(x), y=sum(y)))
}
\seealso{
\code{\link{\%>\%}}, \code{\link{\%<>\%}}, \code{\link{\%T>\%}}, \code{\link{\%$\%}}
}

