% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromNamespace_magrittr.R
\name{debug_fseq}
\alias{debug_fseq}
\title{debug_fseq}
\usage{
debug_fseq(fseq, ...)
}
\arguments{
\item{fseq}{a functional sequence.}

\item{...}{indices of functions to debug.}
}
\description{
Reexported from \code{\link[magrittr]{debug_fseq}}
}
