% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dump_pipes.R
\name{dump_pipes}
\alias{dump_pipes}
\title{Dump the evaluation environments of a pipelines}
\usage{
dump_pipes(dumpto = "last.dump", to.file = FALSE)
}
\arguments{
\item{dumpto}{The name of the object or file to dump to.}

\item{to.file}{Whether to dump to a file. If \code{FALSE},
  then the dump does into an object in the global workspace.}
}
\description{
This function is the equivalent of \code{utils::dump.frames},
for magrittr pipes.
}
\details{
It dumps the calls and the evaluation environments to
an object in the global environment (\code{.GlobalEnv}).
It also adds an attribute called \code{pipes} that contains
information about the pipe stages, for easier debugging.
}
\seealso{
Other pipe.debuggers: \code{\link{print.dump_pipes}};
  \code{\link{tamper}}
}

