% Generated by roxygen2 (4.0.1): do not edit by hand
\name{\%$\%}
\alias{\%$\%}
\title{Expose the columns for direct use in an expression.}
\usage{
lhs \%$\% rhs
}
\arguments{
\item{lhs}{a data.frame or list.}

\item{rhs}{a call/expression.}
}
\description{
This operator exposes columns/members the data.frame/list to the
expression on the right-hand side. This is useful when functions
do not have a data argument, i.e. there is no need to write
\code{data$var} to access \code{var}.
}

