% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{faq-pipe-gender}
\alias{faq-pipe-gender}
\title{FAQ: What is the gender of the pipe?}
\description{
In Magritte's original quote "Ceci n'est pas une pipe," the word
"pipe" is feminine. However the magrittr package quotes it as
"Ceci n'est pas un pipe," with a masculine "pipe." This
lighthearted misappropriation is intentional. Whereas the
object represented in Magritte's painting (a pipe that you can
smoke) is feminine in the French language, a computer pipe
(which is an Anglicism in French) is masculine.
}
