% Generated by roxygen2 (4.0.1): do not edit by hand
\name{freduce}
\alias{freduce}
\title{Apply a list of functions in a sequence}
\usage{
freduce(value, fl)
}
\arguments{
\item{value}{initial value.}

\item{fl}{a list of functions.}
}
\value{
The result after applying each function in turn.
}
\description{
Apply a list of functions in a sequence
}
\details{
this function does not return a vector of the same length as
\code{fl}. It applies the first function to \code{value}, the the next
function to the result of the previous, etc.
}

