% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fseq}
\alias{\%,\%}
\alias{fseq}
\title{Functional sequences ala magrittr}
\usage{
fseq(...)

lhs \%,\% rhs
}
\arguments{
\item{...}{one or more expressions compatible with the rhs parameter in a
  call to \code{\%>\%}. If inputs in one expression are to be reused in a
  subsequent expression, use the form \code{x ~ expr} to name the input
  \code{x}.}

\item{lhs}{expression as in \code{fseq}}

\item{rhs}{expression as in \code{fseq}}
}
\value{
a function.

a function.
}
\description{
Build a function using one or more expressions which could serve as
right-hand sides in a call to \code{\%>\%.}

Functional sequence operator.
}
\examples{
mae <- fseq(residuals, abs, mean)
}

