% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromNamespace_magrittr.R
\name{functions}
\alias{functions}
\title{functions}
\usage{
functions(fseq)
}
\arguments{
\item{fseq}{A functional sequence ala magrittr.}
}
\description{
Reexported from \code{\link[magrittr]{functions}}
}
