% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%T$\%}
\alias{\%T$\%}
\title{Exposition Tee Pipe.}
\usage{
lhs \%T$\% rhs
}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\description{
Exposition Tee Pipe.
}
