% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{is.maybe}
\alias{is.maybe}
\title{Is an object a Maybe object?}
\usage{
is.maybe(x)
}
\arguments{
\item{x}{Any R object}
}
\value{
logical
}
\description{
An S3 method to check if an R object is a Maybe object
}
