% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{is.nothing}
\alias{is.nothing}
\title{Is an object Nothing?}
\usage{
is.nothing(x)
}
\arguments{
\item{x}{Any R object}
}
\value{
logical
}
\description{
An S3 method to check if an R object is a Nothing instance
of a Maybe object
}
