% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{just}
\alias{just}
\title{Just something}
\usage{
just(x)
}
\arguments{
\item{x}{Any R object}
}
\value{
A Maybe object that contains \code{x}
}
\description{
Wraps any object in a safe setting. Moreover, \code{just()} implements
the join functionality of a monad, so \code{just(just(a))} will be \code{just(a)}.
}
