% Generated by roxygen2 (4.0.1): do not edit by hand
\name{memorize}
\alias{memorize}
\title{Memorize variable in pipe operators}
\usage{
lhs \%->\% rhs
}
\arguments{
\item{lhs}{a value}

\item{rhs}{a symbol to assign}
}
\value{
The result of evaluting the left-hand side
}
\description{
Use to memorize temporal variable in a chain.
}
\examples{
\dontrun{
library(dplyr)
library(ggplot2)

x <- 1:10
x \%>\% sum \%->\% y \%>\% multiply_by(10)
y

iris \%>\% group_by(Species) \%>\% summarise(mn=mean(Sepal.Length)) \%->\% res \%>\% qplot(data=.,x=Species,y=mn,geom="bar",stat="identity")
res
x <- 1:10
x \%>\% multiply_by(2) \%->\% y \%->\% z

x <- 1:10
x \%>\% sin \%->\% y \%>\% plot
y

}
}

