% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_pipe.R
\name{new_pipe}
\alias{new_pipe}
\title{Create a new pipe operator.}
\usage{
new_pipe(wrap = {     BODY })
}
\arguments{
\item{wrap}{an expression that will be executed after \code{BODY} is replaced by
the code that would have been run by \code{`%>%` and `.` replaced by the
  input}}
}
\description{
Create a new pipe by wrapping new code around what would have been executed
by the standard \code{\code{\%>\%}} pipe.
}
\details{
Printing existing pipes from the \emph{pipes} package will provide examples to
build on.
}
\examples{
\dontrun{
# How `\%T>\%` and `\%$>\%` were defined in this package 
`\%T>\%` <- new_pipe({local(BODY);.})
`\%$\%`  <- new_pipe(with(., BODY)) 
}
# print the number of rows before and after
`\%nrowba>\%` <- new_pipe({
  print(nrow(.))
  res <- BODY
  print(nrow(res))
  res})
iris \%nrowba>\% head
}
