% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/or_default.R
\name{\%||\%}
\alias{\%||\%}
\title{magrittr or default operator}
\usage{
original_value \%||\% fallback_value
}
\arguments{
\item{original_value}{Any datatype.}

\item{fallback_value}{Any datatype.}
}
\description{
Tests if a value is NA or NULL and returns a default value instead.s
}
\details{
The operator only returns the rhs when lhs is NULL or NA.
Vectors of NULL or NA will not returns the rhs but lhs.s
}
\examples{
x <- NA
y <- x \%||\% 10
}

