% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_friendly_additions.R
\name{pif}
\alias{pif}
\title{Pipe friendly conditional operation}
\usage{
pif(x, p, true, false = identity)
}
\arguments{
\item{x}{An object}

\item{p}{A predicate function, a formula describing such a predicate function, or an expression.}

\item{true, false}{Functions to apply to the data, formulas describing such functions, or expressions.}
}
\value{
The output of \code{true} or \code{false}, either as expressions or applied on data as functions
}
\description{
Apply a transformation on the data only if a condition is met,
by default if condition is not met the input is returned unchanged.
}
\details{
Names of columns or list elements are made accessible as variable to all 3
arguments wether they're given as expressions, functions or formulas
}
\examples{
# using functions
iris \%>\% pif(is.data.frame, dim, nrow)
# using formulas
iris \%>\% pif(~is.numeric(Species), ~"numeric :)",~paste(class(Species)[1],":("))
# using expressions
iris \%>\% pif(nrow(.) > 2, head(.,2))
}
