% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{pipe_eager_lexical}
\alias{pipe_eager_lexical}
\alias{pipe_lazy_masking}
\title{Lazy and eager pipes}
\usage{
pipe_eager_lexical(lhs, rhs)

pipe_lazy_masking(lhs, rhs)
}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\description{
Assign these pipe variants to an infix symbol like \verb{\%>\%}.
}
\keyword{internal}
