% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pipe_env}
\alias{pipe_env}
\title{Function to make a piping environment.}
\usage{
pipe_env(parent, compound = NULL)
}
\arguments{
\item{parent}{the parent which to assign to the new environment.}

\item{compound}{a call which will be used as lhs in compound assignment.
       If not used, set to NULL. This is used to pass forward the
       compound call when environments are locked.}
}
\value{
an environment.
}
\description{
This function is used internally by magrittr to construct environments
for evaluation of chain components.
}

