% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{pipes}
\alias{pipes}
\alias{\%>\%}
\alias{\%T>\%}
\alias{\%$\%}
\alias{\%nowarn>\%}
\alias{\%nomsg>\%}
\alias{\%strict>\%}
\alias{\%try>\%}
\alias{\%P>\%}
\alias{\%V>\%}
\alias{\%D>\%}
\alias{\%summary>\%}
\alias{\%glimpse>\%}
\alias{\%skim>\%}
\alias{\%L>\%}
\alias{\%ae>\%}
\alias{\%compare>\%}
\alias{\%quietly>\%}
\title{Pipe operators to enhance magrittr}
\usage{
lhs \%>\% rhs

lhs \%T>\% rhs

lhs \%$\% rhs

lhs \%nowarn>\% rhs

lhs \%nomsg>\% rhs

lhs \%strict>\% rhs

lhs \%try>\% rhs

lhs \%P>\% rhs

lhs \%V>\% rhs

lhs \%D>\% rhs

lhs \%summary>\% rhs

lhs \%glimpse>\% rhs

lhs \%skim>\% rhs

lhs \%L>\% rhs

lhs \%ae>\% rhs

lhs \%compare>\% rhs

lhs \%quietly>\% rhs
}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\description{
Pipe operators to enhance magrittr
}
\examples{
# print a pipe objects
`\%T>\%`

\dontrun{
# debug the chain
iris \%>\% head(2) \%D>\% `[`(4:5)
}

\dontrun{
# View steps of chain in the viewer
iris \%V>\% head(2) \%V>\% `[`(4:5)
}

# Log steps in the console
iris \%L>\% {Sys.sleep(1);head(.,2)} \%L>\% {Sys.sleep(2);.[4:5]}

# Use print or summary on output
iris \%P>\% head(2) \%P>\% `[`(4:5)
iris \%summary>\% head(2) \%summary>\% `[`(4:5)

\dontrun{
# Use glimpse or skim on output
iris \%glimpse>\% head(2) \%glimpse>\% `[`(4:5) 
iris \%skim>\% head(2) \%skim>\% `[`(4:5) 
}

# Try an operation and ignore if failure
iris \%try>\% {stop("FAILURE!")} \%>\% head(2)

# Silence warnings or messages
-1 \%nowarn>\% sqrt
iris \%nomsg>\% {message("allo?");head(.,2)}

# Fail on a warning
\dontrun{
-1 \%strict>\% sqrt
}


}
