% Generated by roxygen2 (4.0.0): do not edit by hand
\name{prepare}
\alias{prepare}
\title{Function to prepare a block of expressions for pipe chains.}
\usage{
prepare(expr, pipe = "\%>\%")
}
\arguments{
  \item{expr}{A chain of expressions to be prepared for
  later use.}

  \item{pipe}{The pipe to be used. Defaults to
  \code{\%>\%}.  Defaults to \code{quote(x)}.}
}
\value{
A function composed by the arguments.
}
\description{
Alternative to \code{defer}. Very experimental.
This is useful when a sequence of expressions, say \code{a \%>\% b \%>\% c},
contains a subsequence, say \code{b \%>\% c}, which is frequently used.
\code{defer} allows you to compose this subsequence into a singleton which
can be used in subsequent pipe-lines.
}
\examples{
ilm <- prepare(lm(iris) \\\%>\\\% summary)

ilm(Sepal.Length ~ .)
ilm(Sepal.Length ~ . - Species)

ilm <- prepare(lm(Sepal.Length ~ ., data = .) \%>\% summary)
ilm(iris)

verbose_na.omit <-
  prepare(lambda(cat("No. of NAs removed.", x \%>\% is.na \%>\% sum, "\\n\\n")) \%>\%
          na.omit, "\%T>\%")

iris[1:10, 1] <- NA

iris \%>\% verbose_na.omit \%>\% head
}

