% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dump_pipes.R
\name{print.dump_pipes}
\alias{print.dump_pipes}
\title{Pretty-print a pipeline dump}
\usage{
\method{print}{dump_pipes}(x, ...)
}
\arguments{
\item{x}{Object created by \code{\link{dump_pipes}}.}

\item{...}{Ignored.}
}
\value{
The object being printed, invisibly.
}
\description{
Pretty-print a pipeline dump
}
\seealso{
Other pipe.debuggers: \code{\link{dump_pipes}};
  \code{\link{tamper}}
}

