% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromNamespace_magrittr.R
\name{print.fseq}
\alias{print.fseq}
\title{print.fseq}
\usage{
\method{print}{fseq}(x, ...)
}
\arguments{
\item{x}{A functional sequence object}

\item{...}{not used.}
}
\description{
Reexported from \code{\link[magrittr]{print.fseq}}
}
