% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_pipe.R
\name{print.pipe}
\alias{print.pipe}
\title{Print pipe operator}
\usage{
\method{print}{pipe}(x, ...)
}
\arguments{
\item{x}{operator}

\item{...}{not used, for compatibility with other methods}
}
\description{
Method for friendly printing of pipe operators, traditional output can be
obtained by calling \code{print.function(x)}
}
