% Generated by roxygen2 (4.0.1): do not edit by hand
\name{side_effects}
\alias{echo}
\alias{side_effects}
\title{Side-effect functions}
\usage{
... \%>\% echo(...) \%>\% ...
}
\description{
These functions are designed to be used for \emph{specific} side effects.
}
\section{\code{echo}}{


Use \code{echo} for interactive data analysis to print out your intermediate
(or final) results, or for printing out information to log files if your code
is being called in an automated process.
}
\examples{
library(dplyr)
new_iris <- iris \%>\%
  mutate(Log.Sepal.Length = log(Sepal.Length)) \%>\%
  echo(head, str, summary)

# using `\%>\%` within `echo`
library(ggvis)
sepal <- iris \%>\%
   echo(
     ggvis(~Sepal.Length, ~Sepal.Width) \%>\%
       layer_points
   ) \%>\%
   select(Sepal.Length, Sepal.Width)
}

