% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromNamespace_magrittr.R
\name{[.fseq}
\alias{[.fseq}
\title{\code{[.fseq}}
\usage{
\method{[}{fseq}(x, ...)
}
\arguments{
\item{x}{A functional sequence}

\item{...}{index/indices. For double brackets, the index must be of length 1.}
}
\description{
Reexported from \code{\link[magrittr]{[.fseq}}
}
