% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromNamespace_magrittr.R
\name{undebug_fseq}
\alias{undebug_fseq}
\title{undebug_fseq}
\usage{
undebug_fseq(fseq)
}
\arguments{
\item{fseq}{a functional sequence.}
}
\description{
Reexported from \code{\link[magrittr]{debug_fseq}}
}
