% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{unwrap}
\alias{unwrap}
\alias{unwrap.default}
\alias{unwrap.maybe}
\title{Unwrap values from Maybe monad context}
\usage{
unwrap(x, ...)

\method{unwrap}{default}(x, ...)

\method{unwrap}{maybe}(x, ...)
}
\description{
Unwrap values from Maybe monad context
}
