% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magrittr_modified.R
\name{wrap_function}
\alias{wrap_function}
\title{Wrap an expression in a function}
\usage{
wrap_function(body, pipe, env)
}
\arguments{
\item{body}{an expression which will serve as function body in single-argument
function with an argument names \code{.} (a dot)}

\item{pipe}{a quoted magrittr pipe, which determines how the function is made.}

\item{env}{The environment in which to contruct the function.}
}
\value{
a function of a single argument, named \code{.}.
}
\description{
This function takes the "body" part of a function and wraps it in
a function.
}
