% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.fseq}
\alias{print.fseq}
\title{Print method for functional sequences.}
\usage{
\method{print}{fseq}(x, ...)
}
\arguments{
\item{x}{a composite function}

\item{...}{not used.}
}
\description{
Generic method for printing of functional sequences generated with
either \code{\%,\%} or \code{fseq}.
}

