% Generated by roxygen2 (4.0.1): do not edit by hand
\name{:=}
\alias{:=}
\title{Compound assignment operator for magrittr chains.}
\usage{
lhs := rhs
}
\arguments{
\item{lhs}{a reference to a value}

\item{rhs}{a chain expression.}
}
\description{
The \code{`:=`} operator is short-hand notation for altering a value and binding
the existing name to the new value, e.g. rather than \code{a <- a \%>\% b}
one can use \code{a := b}.
}
\examples{
\dontrun{
tmp <- iris

tmp :=
 subset(Species == "setosa") \%>\%
 set_names(LETTERS[1:5])

tmp$A := add(2)

x <- 1:10

x :=
 replace(1:5, 0) \%>\%
 divide_by(2)

}
}

