% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_pipe.R
\name{debug_fseq}
\alias{debug_fseq}
\alias{undebug_fseq}
\title{Debugging function for functional sequences.}
\usage{
debug_fseq(fseq, ...)

undebug_fseq(fseq)
}
\arguments{
\item{fseq}{a functional sequence.}

\item{...}{indices of functions to debug.}
}
\value{
\code{invisible(NULL)}.
}
\description{
This is a utility function for marking functions in a functional
sequence for debbuging.
}
