% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_pipe.R
\name{debug_pipe}
\alias{debug_pipe}
\title{Debugging function for magrittr pipelines.}
\usage{
debug_pipe(x)
}
\arguments{
\item{x}{a value}
}
\value{
x
}
\description{
This function is a wrapper around \code{browser}, which makes it
easier to debug at certain places in a magrittr pipe chain.
}
