% Generated by roxygen2 (4.0.0): do not edit by hand
\name{defer}
\alias{defer}
\title{Compose a chain block of expressions which can be deferred and re-used.}
\usage{
defer(..., arg = quote(x))
}
\arguments{
  \item{...}{Expressions to be composed together with
  \code{\%>\%}}

  \item{arg}{A name for the argument in the resulting
  function.  Defaults to \code{quote(x)}.}
}
\value{
A function composed by the arguments.
}
\description{
This is useful when a sequence of expressions, say \code{a \%>\% b \%>\% c},
contains a subsequence, say \code{b \%>\% c}, which is frequently used.
\code{defer} allows you to compose this subsequence into a singleton which
can be used in subsequent pipe-lines.
}
\examples{
ilm <- defer(lm(iris), summary)

iris \%>\%
  lambda(rbind(z \%>\% head, z \%>\% tail), z)

1:10 \%>\%
  sin \%>\%
  lambda({ d <- abs(x) > 0.5; x*d })
}

