% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freduce.R
\name{freduce}
\alias{freduce}
\title{Apply a list of functions sequentially}
\usage{
freduce(value, function_list)
}
\arguments{
\item{value}{initial value.}

\item{function_list}{a list of functions.}
}
\value{
The result after applying each function in turn.
}
\description{
This function applies the first function to \code{value}, then the
next function to the result of the previous function call, etc.
}
