% Generated by roxygen2 (4.0.1): do not edit by hand
\name{[[.fseq}
\alias{[.fseq}
\alias{[[.fseq}
\title{Extract function(s) from a functional sequence.}
\usage{
\method{[[}{fseq}(x, ...)

\method{[}{fseq}(x, ...)
}
\arguments{
\item{x}{A functional sequence}

\item{...}{index/indices.}
}
\value{
A function.
}
\description{
Extract function(s) from a functional sequence.
}

