% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{[[.fseq}
\alias{[[.fseq}
\alias{[.fseq}
\title{Extract function(s) from a functional sequence.}
\usage{
\method{[[}{fseq}(x, ...)

\method{[}{fseq}(x, ...)
}
\arguments{
\item{x}{A functional sequence}

\item{...}{index/indices. For double brackets, the index must be of length 1.}
}
\value{
A function or functional sequence.
}
\description{
Functional sequences can be subset using single or double brackets.
A single-bracket subset results in a new functional sequence, and
a double-bracket subset results in a single function.
}
