% Generated by roxygen2 (4.0.1): do not edit by hand
\name{functions}
\alias{functions}
\title{Extract the function list from a functional sequence.}
\usage{
functions(fseq)
}
\arguments{
\item{fseq}{A functional sequence ala magrittr.}
}
\value{
a list of functions (of class \code{flist})
}
\description{
Extract the function list from a functional sequence.
}

