% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{functions}
\alias{functions}
\title{Extract the function list from a functional sequence.}
\usage{
functions(fseq)
}
\arguments{
\item{fseq}{A functional sequence ala magrittr.}
}
\value{
a list of functions
}
\description{
This can be used to extract the list of functions inside a functional
sequence created with a chain like \code{. \%>\% foo \%>\% bar}.
}
