% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_pipe.R
\name{is_pipe}
\alias{is_pipe}
\title{Check whether a symbol is a valid magrittr pipe.}
\usage{
is_pipe(pipe)
}
\arguments{
\item{pipe}{A quoted symbol}
}
\value{
logical - TRUE if a valid magrittr pipe, FALSE otherwise.
}
\description{
Check whether a symbol is a valid magrittr pipe.
}
