% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lambda}
\alias{l}
\alias{lambda}
\title{Shorthand notation for anonymous/lambda functions
in magrittr pipelines.}
\usage{
lambda(expr)

l(expr)
}
\arguments{
  \item{expr}{A special kind of expression for the
  anonymous function.  The syntax is \code{symbol ~
  expression}, see the examples.}
}
\value{
a function.
}
\description{
This is an alternative syntax for generating anonymous functions.
When used in chains, the call should be enclosed in parentheses to
force evaluation of the function generation before the left-hand side
is inserted.
}
\details{
\code{lambda} has a special syntax, where the expression is defined as
\code{symbol ~ expression}. The alias \code{l} is shorthand for \code{lambda}.
Previous versions used symbol -> expression syntax, but this caused
problems with compiling packages. There is currently a warning if the
old syntax is used.
}
\examples{
lambda(x ~ x^2 + 2*x)

sapply(1:10, lambda(x ~ x^2))

Filter(lambda(x ~ x > 0), rnorm(100))

iris \%>\%
  (lambda(dfr ~ rbind(dfr \%>\% head, dfr \%>\% tail)))

1:10 \%>\%
  sin \%>\%
  (lambda(x ~ {
    d <- abs(x) > 0.5
    x*d
  }))
}

