% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pipe}
\alias{pipe}
\title{Function to create a pipe operator.}
\usage{
pipe(tee = FALSE)
}
\arguments{
  \item{tee}{logical indicating whether the left-hand side
  should be returned rather than the result of the piped
  expression.}
}
\value{
When tee is FALSE a standard pipe operator is returned. When tee is
  TRUE, a tee operator is returned, i.e. an operator for which
  the right-hand side is used for the side-effect and the left-hand side
  is returned.
}
\description{
Function to create a pipe operator.
}
\details{
This is not exported.
}

