% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{magrittr}
\alias{magrittr}
\alias{magrittr-package}
\title{magrittr - a forward-pipe operator for R}
\description{
This package provides pipe-forwarding mechanisms similar to
(but not exactly like) e.g. F#'s pipe-forward operator. The semantics of
magrittr yield clean and readable code which can be read from left to right,
rather than from the inside and out (which is the case with nested function
calls), and reduce the need for temporary value bindings. This It is
particularly useful for manipulating data frames etc, where multiple steps
are needed before arriving at the desired result. The package also contains
useful functions/aliases which fit well into the syntax promoted by the
package. For a brief introduction, see \code{vignette("magrittr")}.
}
\details{
\tabular{ll}{
 Package: \tab magrittr            \cr
 Type:    \tab Package             \cr
 Version: \tab 1.0.1               \cr
 Date:    \tab 2014-01-14          \cr
 License: \tab MIT + file LICENCE. \cr
}
The main feature is provided by the \code{\%>\%} binary operator. It takes a
value (e.g. a data.frame) on the left-hand side and a function or expression
on the right-hand side, see \code{\link{\%>\%}}.
}
\author{
Stefan Milton Bache and Hadley Wickham.

Maintainer: Stefan Holst Milton Bache <stefan[at]stefanbache.dk>
}

