% Generated by roxygen2 (4.0.1): do not edit by hand
\name{\%>\%}
\alias{\%<>\%}
\alias{\%>\%}
\alias{\%T>\%}
\title{magrittr pipe}
\usage{
lhs \%>\% rhs

lhs \%<>\% rhs

lhs \%T>\% rhs
}
\arguments{
\item{lhs}{either a value or the magrittr placeholder.}

\item{rhs}{a function call using the magrittr semantics.}
}
\value{
value of piping the left-hand side into the right-hand side call.
}
\description{
magrittr pipe
}
\details{
When the magrittr placeholder, \code{.}, is used as \code{lhs}
The result will be a functional sequence, i.e. a function which applies
the entire chain of right-hand sides in turn to its input.
}

