% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pipe_env}
\alias{pipe_env}
\title{Function to make a piping environment.}
\usage{
pipe_env(parent, compound = NULL)
}
\arguments{
  \item{parent}{the parent which to assign to the new
  environment.}

  \item{compound}{a call which will be used as lhs in
  compound assignment.  If not used, set to NULL.}
}
\value{
an environment.
}
\description{
Function to make a piping environment.
}

