% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe2.R
\name{pipeops}
\alias{pipeops}
\alias{\%nowarn>\%}
\alias{\%nomsg>\%}
\alias{\%strict>\%}
\alias{\%try>\%}
\alias{\%P>\%}
\alias{\%L>\%}
\alias{\%V>\%}
\alias{\%D>\%}
\alias{\%summary>\%}
\alias{\%glimpse>\%}
\alias{\%skim>\%}
\alias{\%C>\%}
\alias{\%ae>\%}
\alias{\%compare>\%}
\title{Pipe operators to enhance magrittr}
\usage{
lhs \%nowarn>\% rhs

lhs \%nomsg>\% rhs

lhs \%strict>\% rhs

lhs \%try>\% rhs

lhs \%P>\% rhs

lhs \%L>\% rhs

lhs \%V>\% rhs

lhs \%D>\% rhs

lhs \%summary>\% rhs

lhs \%glimpse>\% rhs

lhs \%skim>\% rhs

lhs \%C>\% rhs

lhs \%ae>\% rhs

lhs \%compare>\% rhs
}
\description{
These operators return the same as magrittr's standard pipe operator but have
additional side effects.
}
\details{
\describe{
\item{\%D>\%}{debugs the pipe chain at the relevant step}
\item{\%V>\%}{uses \code{View()} on the output}
\item{\%L>\%}{logs the relevant call to the console and time it}
\item{\%P>\%}{uses \code{print()} on the output}
\item{\%summary>\%}{prints the \code{summary()} off the output}
\item{\%glimpse>\%}{uses \code{tibble::glimpse} on the output}
\item{\%skim>\%}{uses \code{skimr::skim} on the output}
\item{\%C>\%}{clocks the relevant step}
\item{\%nowarn>\%}{silence warnings}
\item{\%nomsg>\%}{silence messages}
\item{\%strict>\%}{fail on warning}
\item{\%try>\%}{tries, if failure prints error and returns input}
}
}
\examples{
# silence a warning
data.frame(a= c(1,-1)) \%nowarn>\% transform(a=sqrt(a))

# view stepps of chain in the viewer
iris \%V>\% head(2) \%V>\% `[`(4:5)

# log steps in the console
iris \%L>\% {Sys.sleep(1);head(.,2)} \%L>\% {Sys.sleep(2);.[4:5]}

# use print, summary or glimpse on output
iris \%P>\% head(2) \%P>\% `[`(4:5)
iris \%S>\% head(2) \%S>\% `[`(4:5)
iris \%G>\% head(2) \%G>\% `[`(4:5)

# debug the chain
iris \%>\% head(2) \%D>\% `[`(4:5)
}
