% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.fseq}
\alias{print.fseq}
\title{Print method for functional sequence.}
\usage{
\method{print}{fseq}(x, ...)
}
\arguments{
\item{x}{A functional sequence object}

\item{...}{not used.}
}
\value{
x
}
\description{
Print method for functional sequence.
}
