% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%T>\%}
\alias{\%T>\%}
\title{Tee pipe}
\usage{
lhs \%T>\% rhs
}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\description{
Pipe a value forward into a function- or call expression and return the
original value instead of the result. This is useful when an expression
is used for its side-effect, say plotting or printing.
}
\details{
The tee pipe works like \code{\link{\%>\%}}, except the
return value is \code{lhs} itself, and not the result of \code{rhs} function/expression.
}
\examples{
rnorm(200) \%>\%
matrix(ncol = 2) \%T>\%
plot \%>\% # plot usually does not return anything. 
colSums

}
\seealso{
\code{\link{\%>\%}}, \code{\link{\%<>\%}}, \code{\link{\%$\%}}
}
